requirements.txt
apache-airflow==2.10.5
apache-airflow-providers-amazon==9.6.0
Pillow==10.0.0
requests==2.31.0
boto3==1.37.36 
python-dotenv==1.0.0